update bil agar bisa di edit

+++ database/migrations/2023_06_28_121205_create_bills_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             // $table->foreignUuid(column: 'user_id')->constrained(column: 'id')->on('users')->onDelete('cascade');
             $table->string('type', 255);
             $table->date('date_bill');
-            $table->decimal('nominal', 10);
+            $table->integer('nominal');
             $table->timestamps();
         });
     }
